+++ config/dev.exs
@@ -45,7 +45,5 @@ config :td_auth, TdAuth.Repo,
   hostname: "localhost",
   pool_size: 10
 
-config :td_auth, acl_removement: false
-
 # Redis configuration
-config :td_perms, redis_host: "localhost"
+config :td_cache, redis_host: "localhost"