+++ lib/td_auth/accounts/user_acl.ex
@@ -7,7 +7,7 @@ defmodule TdAuth.Accounts.UserAcl do
   alias TdAuth.Permissions.AclEntry
   alias TdAuth.Permissions.Role
   alias TdAuth.Repo
-  alias TdPerms.TaxonomyCache
+  alias TdCache.TaxonomyCache
 
   def get_user_acls(%User{} = user) do
     user
@@ -16,20 +16,19 @@ defmodule TdAuth.Accounts.UserAcl do
   end
 
   defp get_acl_entries(user) do
-    gids = user
-    |> Repo.preload(:groups)
-    |> Map.get(:groups)
-    |> Enum.reduce([], &([&1.id|&2]))
+    gids =
+      user
+      |> Repo.preload(:groups)
+      |> Map.get(:groups)
+      |> Enum.reduce([], &[&1.id | &2])
 
     %{user_id: user.id, gids: gids}
-    |> AclEntry.list_acl_entries_by_user_with_groups
+    |> AclEntry.list_acl_entries_by_user_with_groups()
   end
 
   defp add_to_user_acl(user_acl, %{id: id, type: type}) do
     user_acl
-    |> Map.put(:resource, %{id: id,
-                            type: type,
-                            name: TaxonomyCache.get_name(id)})
+    |> Map.put(:resource, %{id: id, type: type, name: TaxonomyCache.get_name(id)})
   end
 
   defp add_to_user_acl(user_acl, %Role{} = role) do
@@ -42,46 +41,53 @@ defmodule TdAuth.Accounts.UserAcl do
     |> Map.put(:group, Map.take(group, [:id, :name]))
   end
 
-  defp get_user_acl(%AclEntry{resource_type: resource_type,
-                                  resource_id: resource_id,
-                                  principal_type: "user",
-                                  role_id: role_id})
-                                  when resource_type == "domain" do
+  defp get_user_acl(%AclEntry{
+         resource_type: resource_type,
+         resource_id: resource_id,
+         principal_type: "user",
+         role_id: role_id
+       })
+       when resource_type == "domain" do
     case Role.get_role(role_id) do
       role when not is_nil(role) ->
         %{}
         |> add_to_user_acl(%{id: resource_id, type: resource_type})
         |> add_to_user_acl(role)
 
-      _ -> nil
+      _ ->
+        nil
     end
   end
-  defp get_user_acl(%AclEntry{resource_type: resource_type,
-                              resource_id: resource_id,
-                              principal_type: "group",
-                              principal_id: principal_id,
-                              role_id: role_id})
-                              when resource_type == "domain" do
 
+  defp get_user_acl(%AclEntry{
+         resource_type: resource_type,
+         resource_id: resource_id,
+         principal_type: "group",
+         principal_id: principal_id,
+         role_id: role_id
+       })
+       when resource_type == "domain" do
     case {Role.get_role(role_id), Accounts.get_group(principal_id)} do
       {role, group} when not is_nil(role) and not is_nil(group) ->
         %{}
         |> add_to_user_acl(%{id: resource_id, type: resource_type})
         |> add_to_user_acl(role)
         |> add_to_user_acl(group)
-      _ -> nil
+
+      _ ->
+        nil
     end
   end
+
   defp get_user_acl(_), do: nil
 
   defp do_get_user_acls(acl_entries) do
     acl_entries
-    |> Enum.reduce([], fn(acl_entry, acc) ->
+    |> Enum.reduce([], fn acl_entry, acc ->
       case get_user_acl(acl_entry) do
         nil -> acc
-        user_acl -> [user_acl|acc]
+        user_acl -> [user_acl | acc]
       end
     end)
   end
-
 end