+++ lib/td_auth_web/auth_provider/oidc.ex
@@ -3,12 +3,12 @@ defmodule TdAuthWeb.AuthProvider.OIDC do
   Authentication provider for OpenID Connect
   """
 
-  alias TdPerms.NonceCache
+  alias TdCache.NonceCache
   require Logger
 
   def authentication_url do
-    nonce = NonceCache.create_nonce
-    state = NonceCache.create_nonce
+    nonce = NonceCache.create_nonce()
+    state = NonceCache.create_nonce()
 
     OpenIDConnect.authorization_uri(:oidc, %{
       "response_mode" => "fragment",